Refactor spell effect application into modular handlers
